arm: Introduce CPU parking page
This prepares ARM to park CPUs in guest mode instead of host mode. It
will allow to reuse many of the control logic of x86 for managing CPUs
that are in emulated shutdown or in waiting state before a cell starts.

In fact, we could make this a generic pattern later on, converting also
the Intel-specific version to this.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>